gcc_comp_m2_v2/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Vector Database
pinecone-client==2.2.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Authentication & Authorization
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OPA Integration
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
jupyter==1.0.0
ipykernel==6.26.0

# Logging & Monitoring
structlog==23.2.0
